requirements.txt
requests
pandas
kafka-python
pymongo
psycopg2